+++ android/settings.gradle
@@ -1,3 +1,5 @@
 rootProject.name = 'Visible'
+include ':react-native-navigation'
+project(':react-native-navigation').projectDir = new File(rootProject.projectDir, '../node_modules/react-native-navigation/android/app')
 
 include ':app'
