+++ package.json
@@ -9,7 +9,8 @@
   },
   "dependencies": {
     "react": "16.8.3",
-    "react-native": "0.59.8"
+    "react-native": "0.59.8",
+    "react-native-navigation": "^1.1.493"
   },
   "devDependencies": {
     "@babel/core": "^7.4.5",
